Build icon content uri from generated playlist icon filename
This also deletes icons along their playlist.

+++ core-ui/src/main/java/fr/nihilus/music/core/ui/actions/ManagePlaylistAction.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Thibault Seisel
+ * Copyright 2021 Thibault Seisel
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,20 +29,17 @@ import fr.nihilus.music.core.database.playlists.PlaylistTrack
 import fr.nihilus.music.core.media.MediaId
 import fr.nihilus.music.core.media.MediaId.Builder.TYPE_PLAYLISTS
 import fr.nihilus.music.core.os.Clock
-import fr.nihilus.music.core.os.FileSystem
+import fr.nihilus.music.core.os.IconContentUri
+import fr.nihilus.music.core.os.PlaylistIconDir
 import kotlinx.coroutines.withContext
+import java.io.File
 import javax.inject.Inject
+import javax.inject.Provider
 
 /**
  * The width/height in pixels of the generated playlist icons.
  */
 private const val ICON_SIZE_PX = 320
-
-/**
- * The name of the folder into which generated playlist icons are written.
- * This folder only contains PNG files.
- */
-private const val PLAYLIST_ICON_FOLDER = "playlist_icons"
 
 /**
  * Handler for actions related to the management of user-defined playlists.
@@ -57,13 +54,15 @@ private const val PLAYLIST_ICON_FOLDER = "playlist_icons"
  * but should have a non-`null` [track identifier][MediaId.track].
  *
  * @property playlistDao Handle to the playlists storage.
- * @property files Handle to read/write internal storage files.
+ * @property iconDir Directory into which generated playlist icons are stored.
+ * @property baseIconUri Base content uri from which playlist icons could be resolved.
  * @property clock Provider of the system time.
  * @property dispatchers Set of coroutine dispatchers.
  */
 class ManagePlaylistAction @Inject constructor(
     private val playlistDao: PlaylistDao,
-    private val files: FileSystem,
+    @PlaylistIconDir private val iconDir: Provider<File>,
+    @IconContentUri private val baseIconUri: Uri,
     private val clock: Clock,
     private val dispatchers: AppDispatchers
 ) {
@@ -115,7 +114,15 @@ class ManagePlaylistAction @Inject constructor(
      */
     suspend fun deletePlaylist(targetPlaylist: MediaId) {
         val playlistId = requirePlaylistId(targetPlaylist)
-        playlistDao.deletePlaylist(playlistId)
+        val playlist = playlistDao.findPlaylist(playlistId)
+        if (playlist != null) {
+            val iconFilename = playlist.iconUri?.lastPathSegment
+            if (iconFilename != null) withContext(dispatchers.IO) {
+                val iconFile = File(iconDir.get(), iconFilename)
+                iconFile.delete()
+            }
+            playlistDao.deletePlaylist(playlistId)
+        }
     }
 
     private suspend fun generatePlaylistIcon(playlistName: String): Uri? = withContext(dispatchers.Default) {
@@ -126,14 +133,24 @@ class ManagePlaylistAction @Inject constructor(
             )
         }
 
-        val outputBitmap = Bitmap.createBitmap(ICON_SIZE_PX, ICON_SIZE_PX, Bitmap.Config.ARGB_8888)
-        iconSpec.drawToBitmap(outputBitmap)
+        val iconBitmap = Bitmap.createBitmap(ICON_SIZE_PX, ICON_SIZE_PX, Bitmap.Config.ARGB_8888)
+        iconSpec.drawToBitmap(iconBitmap)
+
+        // Sanitize playlist name to make a filename without spaces
+        val iconFilename = playlistName.replace(reWhitespaces, "_") + ".png"
 
         withContext(dispatchers.IO) {
-            // Sanitize playlist name to make a filename without spaces
-            val fileName = playlistName.replace(reWhitespaces, "_")
-            files.writeBitmapToInternalStorage("$PLAYLIST_ICON_FOLDER/$fileName.png", outputBitmap)
+            val iconDir = iconDir.get()
+            check(iconDir.exists() || iconDir.mkdirs()) {
+                "Unable to create playlist icon directory"
+            }
+            val iconFile = File(iconDir, iconFilename)
+            iconFile.outputStream().use {
+                iconBitmap.compress(Bitmap.CompressFormat.PNG, 100, it)
+            }
         }
+
+        Uri.withAppendedPath(baseIconUri, iconFilename)
     }
 
     private fun requirePlaylistId(playlist: MediaId): Long {
